redesigned admin dash to encompass add & view products, edit products separates body from image
