+++ server/controllers/sneakers.controller.js
@@ -56,65 +56,42 @@ module.exports = {
             })
 
     },
-
-    updateSneaker: async (req, res) => {
-        console.log(req.body)
-        try {
-            console.log(req.body, "...", req.file)
-            const { name, brand, gender, price, discountedPrice, color, size, description } = req.body
-            const image = req.file ? req.file.filename : null
-
-            const updatedSneaker = await SneakerFinder.findOneAndUpdate(
-                { _id: req.params.id },
-                { name, brand, gender, price, discountedPrice, color, size, description, image },
-                { new: true }
-            )
-
-            res.status(201).json({ message: 'Sneaker updated successfully!', sneaker: updatedSneaker })
-            console.log(updatedSneaker)
-        } catch (error) {
-            console.log("Something went wrong updating a shoe", error)
-            res.status(500).json({ message: 'Error updating sneaker', error: error })
-        }
-    },
     updateSneakerBody: async (req, res) => {
         console.log(req.body)
         try {
             console.log(req.body, "...", req.file)
             const { name, brand, gender, price, discountedPrice, color, size, description } = req.body
-            const image = req.file ? req.file.filename : null
 
             const updatedSneaker = await SneakerFinder.findOneAndUpdate(
                 { _id: req.params.id },
-                { name, brand, gender, price, discountedPrice, color, size, description, image },
+                { name, brand, gender, price, discountedPrice, color, size, description },
                 { new: true }
             )
 
-            res.status(201).json({ message: 'Sneaker updated successfully!', sneaker: updatedSneaker })
-            console.log(updatedSneaker)
+            res.status(201).json({ message: 'Sneaker body updated successfully!', sneaker: updatedSneaker })
+            console.log("72", updatedSneaker)
         } catch (error) {
-            console.log("Something went wrong updating a shoe", error)
-            res.status(500).json({ message: 'Error updating sneaker', error: error })
+            console.log("74 Something went wrong updating a shoe body", error)
+            res.status(500).json({ message: '75 Error updating sneaker body', error: error })
         }
     },
     updateSneakerImage: async (req, res) => {
         console.log(req.body)
         try {
             console.log(req.body, "...", req.file)
-            const { name, brand, gender, price, discountedPrice, color, size, description } = req.body
             const image = req.file ? req.file.filename : null
 
             const updatedSneaker = await SneakerFinder.findOneAndUpdate(
                 { _id: req.params.id },
-                { name, brand, gender, price, discountedPrice, color, size, description, image },
+                { image },
                 { new: true }
             )
 
-            res.status(201).json({ message: 'Sneaker updated successfully!', sneaker: updatedSneaker })
+            res.status(201).json({ message: 'Sneaker image updated successfully!', sneaker: updatedSneaker })
             console.log(updatedSneaker)
         } catch (error) {
-            console.log("Something went wrong updating a shoe", error)
-            res.status(500).json({ message: 'Error updating sneaker', error: error })
+            console.log("Something went wrong updating a shoe image", error)
+            res.status(500).json({ message: 'Error updating sneaker image', error: error })
         }
     },
