+++ server/models/sneakers.model.js
@@ -34,7 +34,7 @@ const SneakerSchema = new mongoose.Schema({
     },
     size: {
         type: Number,
-        required: [true, 'Shoe size is required'],
+        required: [true, 'Shoe size is must be a whole number between 1 and 18, inclusive'],
         enum: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18]
     },
     description: {
