+++ server/routes/sneakers.routes.js
@@ -6,7 +6,8 @@ module.exports = (app) => {
     app.get('/api/shoes', SneakerController.getAllSneakers);
     app.get('/api/shoes/:id', SneakerController.getOneSneaker);
     app.post('/api/shoes', uploadMiddleware.single('image'), SneakerController.createSneaker);
-    app.put('/api/shoes/:id', uploadMiddleware.single('image'), SneakerController.updateSneaker);
+    app.put('/api/shoes/:id', SneakerController.updateSneakerBody);
+    app.put('/api/shoes/:id/image', uploadMiddleware.single('image'), SneakerController.updateSneakerImage);
     // app.patch('/api/shoes/addPicture', SneakerController.addPicture);
     app.delete('/api/shoes/:id', SneakerController.deleteSneaker);
 }
